% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure}
\alias{configure}
\title{Set some personal defaults for skeletor}
\usage{
configure(name = NULL, email = NULL, github = NULL,
  file = "~/.Rprofile")
}
\arguments{
\item{name}{character: your name}

\item{email}{character: your email address}

\item{github}{character: your GitHub account}

\item{file}{character: file path to set these in as options. Default is your
personal .Rprofile file.}
}
\value{
Invisibly, the options set in \code{file}. Typically, you'll call
this function only once, and only for its side effects of adding the values
to your .Rprofile so that \code{\link{skeletor}} will find them there without
your having to enter them each time.
}
\description{
Set some personal defaults for skeletor
}

