% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletor.R
\name{skeletor}
\alias{skeletor}
\title{Create a package skeleton}
\usage{
skeletor(pkg, dir = pkg, name = getOption("skeletor.name"),
  email = getOption("skeletor.email"),
  github = getOption("skeletor.github"))
}
\arguments{
\item{dir}{character path in which to create the package. Default is \code{name}}

\item{name}{character name for your new package}
}
\value{
The path, \code{dir}, invisibly.
}
\description{
Create a package skeleton
}

