% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletor.R
\name{skeletor}
\alias{skeletor}
\title{Create a package skeleton}
\usage{
skeletor(pkg, dir = pkg, name = getOption("skeletor.name"),
  email = getOption("skeletor.email"),
  github = getOption("skeletor.github"), api = FALSE)
}
\arguments{
\item{pkg}{character: name for your new package}

\item{dir}{character: path in which to create the package. Default is \code{pkg}}

\item{name}{character: your name. Default is 'skeletor.name' from options.
See \code{\link[=configure]{configure()}} to set \code{name}, \code{email}, and
\code{github} in your .Rprofile so that you don't need to pass them in here.}

\item{email}{character: your email address. Likewise taken from option
'skeletor.email' by default.}

\item{github}{character: the GitHub account where you will push this new
package. Likewise taken from option 'skeletor.github' by default.}

\item{api}{logical: is this package an API wrapper? If \code{TRUE}, an \code{api.R}
file of boilerplate code will be added, \code{httr} will be added to \code{Imports},
\code{httptest} will be added to \code{Suggests}, and some basic tests of the wrapping
code will be added. Default is \code{FALSE}.}
}
\value{
The path, \code{dir}, invisibly.
}
\description{
Create a package skeleton
}
